refactor(MatchResult): MatchResult 미사용 기능 제거 및 패키지 이동
