+++ src/test/java/model/participant/PlayerTest.java
@@ -7,7 +7,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import model.card.Card;
 import model.card.Emblem;
 import model.card.Number;
-import model.casino.MatchResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
